chore: move __test__ to root directory
